Revert "Merge cypress-accessibility-test"
