--- bin/accessibility-automation/constants.js
@@ -1,1 +1,0 @@
-exports.API_URL = 'https://accessibility.browserstack.com/api';
