+++ bin/commands/runs.js
@@ -30,12 +30,6 @@ const {
   printBuildLink
 } = require('../testObservability/helper/helper');
 
-
-const { 
-  createAccessibilityTestRun,
-  checkAccessibilityPlatform
-} = require('../accessibility-automation/helper');
-
 module.exports = function run(args, rawArgs) {
 
   markBlockStart('preBuild');
@@ -63,8 +57,6 @@ module.exports = function run(args, rawArgs) {
       Set testObservability & browserstackAutomation flags
     */
     const [isTestObservabilitySession, isBrowserstackInfra] = setTestObservabilityFlags(bsConfig);
-    const checkAccessibility = checkAccessibilityPlatform(bsConfig);
-    const isAccessibilitySession = bsConfig.run_settings.accessibility || checkAccessibility;
 
     utils.setUsageReportingFlag(bsConfig, args.disableUsageReporting);
 
@@ -139,10 +131,6 @@ module.exports = function run(args, rawArgs) {
 
       // add cypress dependency if missing
       utils.setCypressNpmDependency(bsConfig);
-      
-      if (isAccessibilitySession && isBrowserstackInfra) {
-        await createAccessibilityTestRun(bsConfig);
-      }
     }
 
     const { packagesInstalled } = !isBrowserstackInfra ? false : await packageInstaller.packageSetupAndInstaller(bsConfig, config.packageDirName, {markBlockStart, markBlockEnd});
